requirements.txt
django-cms==3.4.1
cmsplugin_contact_plus
cmsplugin-filer>=1.1
djangocms-snippet>=1.9
djangocms-text-ckeditor>=3.2.1
